refactor: decouple DocumentEditToolbar from DocumentEdit
