+++ frontend/actions/actionTypes.js
@@ -27,6 +27,7 @@ export const SET_API_STATUS = 'SET_API_STATUS'
 //
 // Documents
 //
+export const ATTEMPT_SAVE_DOCUMENT = 'ATTEMPT_SAVE_DOCUMENT'
 export const DELETE_DOCUMENTS = 'DELETE_DOCUMENTS'
 export const CLEAR_DOCUMENT_LIST = 'CLEAR_DOCUMENT_LIST'
 export const SET_DOCUMENT_SELECTION = 'SET_DOCUMENT_SELECTION'
