+++ frontend/views/DocumentCreateView/DocumentCreateView.jsx
@@ -5,6 +5,7 @@ import {h, Component} from 'preact'
 import {setPageTitle} from 'lib/util'
 
 import DocumentEdit from 'containers/DocumentEdit/DocumentEdit'
+import DocumentEditToolbar from 'containers/DocumentEditToolbar/DocumentEditToolbar'
 import Header from 'containers/Header/Header'
 import Main from 'components/Main/Main'
 import Page from 'components/Page/Page'
@@ -30,6 +31,12 @@ export default class DocumentCreateView extends Component {
             section={section}
           />
         </Main>
+
+        <DocumentEditToolbar
+          collection={collection}
+          group={group}
+          section={section}
+        />        
       </Page>
     )    
   }
